+++ src/Components/organisms/Navbar.tsx
@@ -9,9 +9,9 @@ import {
   NavbarBrand,
 } from "@nextui-org/react";
 import { Link } from "react-router-dom";
-import curriculum from "../assets/curriculum/Curriculum-Smailen-Vargas.pdf";
-import { useGlobalContext } from "../utils/context";
-import useChangeAvatar from "../utils/useChangeAvatar";
+import curriculum from "../../assets/curriculum/Curriculum-Smailen-Vargas.pdf";
+import { useGlobalContext } from "../../utils/context";
+import useChangeAvatar from "../../utils/useChangeAvatar";
 
 const NavbarSample = () => {
   const { currentAvatar, animation } = useChangeAvatar();
