+++ src/screen/HomeScreen.tsx
@@ -1,6 +1,6 @@
-import Footer from "../Components/Footer";
+import Footer from "../Components/organisms/Footer";
 import Hero from "../Components/Hero";
-import NavbarSample from "../Components/Navbar";
+import NavbarSample from "../Components/organisms/Navbar";
 import Presentation from "../Components/Presentation";
 import Skills from "../Components/Skills";
 import Socials from "../Components/Socials";
